Fixed issues on RPi4 32-/64-bit

+++ so3/include/memory.h
@@ -49,7 +49,7 @@ typedef struct {
 
 struct mem_info {
 	addr_t phys_base;
-    uint32_t size;
+    unsigned long size;
     uint32_t avail_pages; /* Available pages including frame table, without the low kernel region */
 };
 typedef struct mem_info mem_info_t;
